requirements.txt
Django==6.0.2
Pillow==12.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg[binary]==3.2.13
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
